refactor: InputProcessor에서 인풋 분석 로직을 InputParser로 분리
InputProcessor.java 에서 인풋을 분석하고 숫자를 뽑아내는 로직을 InputParser.java로 분리
