+++ src/main/java/calculator/Application.java
@@ -4,10 +4,10 @@ import calculator.controller.Calculator;
 
 public class Application {
     public static void main(String[] args) {
-        Calculator controller = new Calculator();
+        Calculator calculator = new Calculator();
 
-        controller.start();
-        controller.process();
-        controller.display();
+        calculator.start();
+        calculator.process();
+        calculator.display();
     }
 }
